Fix : detail 사용자 이름 반영
